AJ MacLeod:
Attached is a diff for the An2 which makes it work properly in MP sessions.

The FDM modifications just provide the standard normalised control surface
position outputs and don't affect the flight characteristics at all.

Leading slashes in the model.xml have been removed, and the standard
properties (created in the FDM update, above) are used to animate the control
surfaces.